refactor(mobile): move widgets to new files in the widget folder
